requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.2.0
qrcode[pil]==7.4.2